+++ CHANGES.rst
@@ -1,6 +1,17 @@
 Changes
 =======
 
+0.2
+---
+**November 30, 2012**
+
+- Fixed bug in ``MockClient`` to pop the correct response.
+- Added ``restorm.conf.settings`` to store the DEFAULT_CLIENT to use.
+- Added ability to create a development server from your ``MockApiClient``.
+- Added ``root_uri`` parameter to ``Client`` constructor.
+- Added initial XML client implemention.
+- Added initial documentation.
+
 0.1
 ---
 **November 9, 2012**